Refactor specs to handle change in units
ffprobe returns its frame rate as a rational number as opposed to a
floating point (e.g., 24600/1469 instead of 16.75). This allows the
number to maintain a higher precision but is somewhat of a breaking
change for the specs and thus the API.

ffprobe's default units are bit/s instead of the old ffmpeg output being
in kb/s. Again, higher precision but a breaking change for the specs and
thus the API.

ffprobe does not provide a property like FFMPEG::Movie#video_stream or
FFMPEG::Movie#audio_stream. Therefore, an attempt to emulate it and
provide backwards compatibility is provided.